test: convert jest to vitest
