+++ biome.json
@@ -6,7 +6,10 @@
     "linter": {
         "enabled": true,
         "rules": {
-            "recommended": true
+            "recommended": true,
+            "complexity": {
+                "useLiteralKeys": "off"
+            }
         }
     },
     "formatter": {
