--- jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-    preset: 'ts-jest',
-    testEnvironment: 'node',
-}
