+++ package.json
@@ -50,24 +50,21 @@
     },
     "devDependencies": {
         "@biomejs/biome": "1.9.4",
-        "@types/jest": "29.5.14",
         "@types/js-yaml": "4.0.9",
         "@types/node": "22.13.5",
+        "@vitest/coverage-v8": "3.0.6",
         "commitlint": "19.7.1",
         "conventional-changelog-cli": "5.0.0",
-        "jest": "29.7.0",
-        "nyc": "17.1.0",
-        "ts-jest": "29.2.5",
-        "typescript": "5.7.3"
+        "typescript": "5.7.3",
+        "vitest": "3.0.6"
     },
     "scripts": {
         "build": "pnpm run build:esm && pnpm run build:cjs",
         "build:esm": "tsc -p tsconfig.lib.json --outDir lib/esm",
         "build:cjs": "tsc -p tsconfig.lib.json --module CommonJS --moduleResolution Node --outDir lib/cjs",
         "postbuild:cjs": "echo '{\"type\": \"commonjs\"}' > lib/cjs/package.json",
-        "test": "jest",
-        "test-cover": "jest --coverage",
-        "coverage": "jest --coverage --coverageReporters=text-lcov | coveralls",
+        "test": "vitest --run",
+        "coverage": "vitest run --coverage",
         "version": "echo ${npm_package_version}",
         "lint": "biome check --write",
         "lint-check": "biome check",
