+++ src/types/global.d.ts
@@ -2,5 +2,6 @@ declare namespace NodeJS {
     interface ProcessEnv {
         CONF_DIR?: string
         CONF_FILES?: string
+        [key: string]: string | number | boolean | undefined
     }
 }
