+++ vitest.config.js
@@ -1,0 +1,21 @@
+import { defineConfig } from 'vitest/config'
+
+export default defineConfig({
+    test: {
+        globals: true,
+        environment: 'node',
+        include: ['test/**/*.ts'],
+        coverage: {
+            reporter: ['lcov', 'cobertura', 'text'],
+            include: ['src/**'],
+            exclude: ['src/types/**'],
+        },
+    },
+    esbuild: {
+        target: 'esnext',
+    },
+    resolve: {
+        extensions: ['.ts', '.js'],
+        conditions: ['import', 'module', 'default'],
+    },
+})
